исправил код к задаче 71

+++ Tasks/Ex071/Program.cs
@@ -2,7 +2,7 @@
 
 Console.WriteLine("71. Написать программу вычисления функции Аккермана A(n, m)");
 
-int n = GetNumber("Введите число n: ");
-int m = GetNumber("Введите число m: ");
+ulong n = GetNumber("Введите число n: ");
+ulong m = GetNumber("Введите число m: ");
 
 Console.WriteLine($"Результат вычисления функции Аккермана A({n}, {m}) = " + A(n, m));
